refactor: MainActivity 내 Composable 함수 파일 분리
