+++ app/src/main/java/woowacourse/payments/MainActivity.kt
@@ -4,38 +4,12 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 
@@ -64,136 +38,32 @@ class MainActivity : ComponentActivity() {
                                 .padding(10.dp)
                                 .align(Alignment.CenterHorizontally)
                         )
-                        CardNumber()
-                        CardExpirationDate()
-                        CardName()
-                        CardPassword()
+                        DigitFieldText(
+                            label = "카드 번호",
+                            hint = "0000 - 0000 - 0000 - 0000",
+                            maxLength = 16,
+                            mask = InputMask.CardNumber,
+                            )
+                        DigitFieldText(label = "만료일",
+                            hint = "MM / YY",
+                            fraction = 0.5f,
+                            maxLength = 4,
+                            mask = InputMask.Expiry,
+                            )
+                        LimitedTextField(
+                            label = "카드 소유자 이름(선택)",
+                            hint = "카드에 표시된 이름을 입력하세요.",
+                            maxLength = 30,
+                        )
+                        DigitFieldText(label = "비밀번호",
+                            hint = "0000",
+                            fraction = 0.5f,
+                            maxLength = 4,
+                            mask = InputMask.Password,
+                            )
                     }
                 }
             }
         }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun NewCardTopBar(
-    onBackClick: () -> Unit,
-    onSaveClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    TopAppBar(
-        title = { Text(text = "카드 추가") },
-        navigationIcon = {
-            IconButton(onClick = { onBackClick() }) {
-                Icon(
-                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                    contentDescription = "뒤로 가기"
-                )
-            }
-        },
-        actions = {
-            IconButton(onClick = { onSaveClick() }) {
-                Icon(
-                    imageVector = Icons.Filled.Check,
-                    contentDescription = "완료"
-                )
-            }
-        },
-        modifier = modifier
-    )
-}
-
-@Composable
-fun PaymentCard(modifier: Modifier = Modifier) {
-    Box(
-        contentAlignment = Alignment.CenterStart,
-        modifier = modifier
-            .size(208.dp, 124.dp)
-            .shadow(8.dp)
-            .background(
-                color = Color(0xFF333333),
-                shape = RoundedCornerShape(5.dp),
-            )
-    ) {
-        Box(
-            modifier = modifier
-                .padding(start = 14.dp, bottom = 10.dp)
-                .size(40.dp, 28.dp)
-                .background(
-                    color = Color(0xFFCBBA64),
-                    shape = RoundedCornerShape(5.dp)
-                )
-        )
-    }
-}
-
-@Composable
-fun CardNumber(modifier: Modifier = Modifier) {
-    var number by remember { mutableStateOf("") }
-    OutlinedTextField(
-        value = number,
-        onValueChange = { number = it },
-        label = { Text(text = "카드 번호") },
-        placeholder = { Text(text = "0000 - 0000 - 0000 - 0000") },
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(horizontal = 24.dp, vertical = 15.dp)
-    )
-}
-
-@Composable
-fun CardExpirationDate(modifier: Modifier = Modifier) {
-    var text by remember { mutableStateOf("") }
-    OutlinedTextField(
-        value = text,
-        onValueChange = { text = it },
-        label = { Text(text = "만료일") },
-        placeholder = { Text(text = "MM / YY") },
-        modifier = modifier
-            .fillMaxWidth(0.5f)
-            .padding(horizontal = 24.dp, vertical = 15.dp)
-    )
-}
-
-@Composable
-fun CardName(modifier: Modifier = Modifier) {
-    var text by remember { mutableStateOf("") }
-    OutlinedTextField(
-        value = text,
-        onValueChange = { text = it },
-        label = { Text(text = "카드 소유자 이름(선택)") },
-        placeholder = { Text(text = "카드에 표시된 이름을 입력하세요.") },
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(horizontal = 24.dp, vertical = 15.dp),
-        supportingText = { Text(
-            text = "${text.length}/30",
-            modifier = modifier.fillMaxWidth(),
-            textAlign = TextAlign.End,
-            ) }
-    )
-}
-
-@Composable
-fun CardPassword(modifier: Modifier = Modifier,
-                 isHide: Boolean = false) {
-    var text by remember { mutableStateOf("") }
-    OutlinedTextField(
-        value = text,
-        onValueChange = { text = it },
-        label = { Text(text = "비밀번호") },
-        placeholder = { Text(text = "0000") },
-        modifier = modifier
-            .fillMaxWidth(0.5f)
-            .padding(horizontal = 24.dp, vertical = 10.dp),
-        visualTransformation = if (isHide) VisualTransformation.None else PasswordVisualTransformation()
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    AndroidpaymentsTheme {
     }
 }
